introduce hooks for control panel
